Search v3 (#22)
* Adds logic to allow `default_params`

This adds a “global” default param set of `%{}` and specifically marks the function as `overridable`.  It also adds some logic to incorporate any default parameters into the provided query string.  The `params` are merged into the defaults to allow the caller full control if they wish to override one of the defaults for a specific call.

* Uses `query_string` function so that defaults are included.

* Adds `default_params` for `User`.

This adds the new `search_engine` parameter to all `User` queries unless explicitly configured not to in which case empty defaults are used instead.

* Explain the config option for `v2` search

* Bumps `httpoison`